requirements.txt
# Core
openai>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
tqdm>=4.60.0

# Media Processing
opencv-python>=4.8.0
numpy>=1.24.0
pydub>=0.25.0

# Web Tools
beautifulsoup4>=4.12.0
lxml>=4.9.0
chardet>=5.0.0

# Code Executor (sandbox environment)
sympy>=1.12
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=14.0.0
h5py>=3.9.0
Pillow>=10.0.0
fpdf2>=2.7.0
reportlab>=4.0.0
PyYAML>=6.0.0

# Optional: ASR (for --use_asr mode)
# openai-whisper>=20231117

